+++ helpers/BaseCrudQuery.php
@@ -9,7 +9,9 @@ abstract class BaseCrudQuery
 {
   public $table;
 
+  protected $preparedTypes = "";
   protected $components = [];
+  protected $variables = [];
   protected $query;
   
   protected static $unionQueries = [];
@@ -33,6 +35,16 @@ abstract class BaseCrudQuery
     return $this->query;
   }
 
+  public function getVariables()
+  {
+    return $this->variables;
+  }
+
+  public function getPreparedTypes()
+  {
+    return $this->preparedTypes;
+  }
+
   public function union(string $table = "")
   {
     $unionQuery = new static();
@@ -48,7 +60,7 @@ abstract class BaseCrudQuery
       $this->endQuery();
     }
 
-    array_push(static::$unionQueries, $this->query);
+    array_push(static::$unionQueries, $this);
     return $unionQuery;
   }
 
@@ -70,9 +82,48 @@ abstract class BaseCrudQuery
   protected function createUnionQuery()
   {
     if(!empty(static::$unionQueries)){
-      array_push(static::$unionQueries, $this->query);
-      $this->query = implode(" UNION ", static::$unionQueries);
+
+      $variables = [];
+      $preparedTypes = "";
+      $unionQueries = [];
+
+      foreach(static::$unionQueries as $query)
+      {
+        array_push($variables, ...$query->getVariables());
+        $preparedTypes .= $query->getPreparedTypes();
+        array_push($unionQueries, $query->getQuery());
+      }
+
+      array_push($variables, ...$this->variables);
+      $this->preparedTypes = $preparedTypes . $this->preparedTypes;
+      array_push($unionQueries, $this->query);
+
+      $this->variables = $variables;
+
+      $this->query = implode(" UNION ", $unionQueries);
       static::$unionQueries = [];
+    }
+  }
+
+  protected function isStringIntDouble(...$variables)
+  {
+    foreach ($variables as $variable) {
+      if(!is_string($variable) && !is_int($variable) && !is_double($variable)){
+        throw new \InvalidArgumentException("Argument has to be of type string, int or double");
+      }
+    }
+  }
+
+  protected function returnTypeStringIntDouble($variable)
+  {
+    if(is_string($variable)){
+      return "s";
+    } else if (is_int($variable)) {
+      return "i";
+    } else if (is_double($variable)) {
+      return "d";
+    } else {
+      throw new \InvalidArgumentException("Argument has to be of type string, int, or double");
     }
   }