+++ helpers/CrudQueryWhereHelper.php
@@ -7,8 +7,12 @@ namespace helpers;
 
 trait CrudQueryWhereHelper
 {
-  public function whereEquals(string $field, string $value, bool $notEquals = false)
+  public function whereEquals(string $field, $value, bool $notEquals = false)
   {
+    $this->isStringIntDouble($value);
+    array_push($this->variables, $value);
+    $this->preparedTypes .= $this->returnTypeStringIntDouble($value);
+
     $firstWhere = false;
     $field = $this->encloseBackticks($field);
     $value = "'$value'";
@@ -20,12 +24,16 @@ trait CrudQueryWhereHelper
     }
 
     ($firstWhere) ? $field = "WHERE $field" : false;
-    array_push($this->components["where"], "$field $condition $value");
+    array_push($this->components["where"], "$field $condition ?");
     return $this;
   }
 
-  public function whereGreaterThan(string $field, string $value, bool $orEqualTo = false)
+  public function whereGreaterThan(string $field, $value, bool $orEqualTo = false)
   {
+    $this->isStringIntDouble($value);
+    array_push($this->variables, $value);
+    $this->preparedTypes .= $this->returnTypeStringIntDouble($value);
+
     $firstWhere = false;
     $condition = ">";
     ($orEqualTo) ? $condition .= "=" : false;
@@ -38,12 +46,16 @@ trait CrudQueryWhereHelper
     
     }
     ($firstWhere) ? $field = "WHERE $field" : false;
-    array_push($this->components["where"], "$field $condition $value");
+    array_push($this->components["where"], "$field $condition ?");
     return $this;
   }
 
-  public function whereLessThan(string $field, string $value, bool $orEqualTo = false)
+  public function whereLessThan(string $field, $value, bool $orEqualTo = false)
   {
+    $this->isStringIntDouble($value);
+    array_push($this->variables, $value);
+    $this->preparedTypes .= $this->returnTypeStringIntDouble($value);
+
     $firstWhere = false;
     $condition = "<";
     ($orEqualTo) ? $condition .= "=" : false;
@@ -56,7 +68,7 @@ trait CrudQueryWhereHelper
     }
     
     ($firstWhere) ? $field = "WHERE $field" : false;
-    array_push($this->components["where"], "$field $condition $value");
+    array_push($this->components["where"], "$field $condition ?");
     return $this;
   }
 }