+++ changelog.md
@@ -18,6 +18,7 @@
 - changed the Query object into seperate Objects per crud action
 - changed the query creators to create better SQL statements
 - changed the enclosedValues function to now be able to be used for non-enclosedValues as well
+- changed the the insert and all where statements into prepared statements
 - fixed a bug where getRowsByFieldValue would always use the rows array instead of the given array
 - fixed a bug where iterate would empty the modifiedRows array, and break the method chain
 - fixed a bug where you couldn't get the results from the excecuteQuery function
@@ -30,6 +31,7 @@
 - renamed folder access => database
 - renamed enclosedValues => multipleValues
 - renamed mutlipleValues => multipleDropValues
+- renamed excecuteQuery => executeQuery
 - removed the QueryEngine class
 - removed the BaseEngine class
 - removed the BaseQuery class